No more elementwise access of vectors (140 ms)

+++ Mandelbrot/Painter.h
@@ -38,10 +38,11 @@ private:
     int pixelWidth;
     int pixelHeight;
     int maxIterations;
-    double *LUT;
-    double *GreenLUT;
-    double *RedLUT;
-    double *BlueLUT;
+    float *LUT;
+    float *GreenLUT;
+    float *RedLUT;
+    float *BlueLUT;
+
 
     // The number of colours in the table, excluding the last one (black)
     // There can be at most 15, since we need a dummy black after the last.
